+++ templates/visuals/rhtml/src/visual.ts
@@ -24,7 +24,7 @@
 *  THE SOFTWARE.
 */
 "use strict";
-import powerbi from "powerbi-visuals-api";
+import powerbi from "@visualbi/powerbi-visuals-api";
 import VisualConstructorOptions = powerbi.extensibility.visual.VisualConstructorOptions;
 import VisualUpdateOptions = powerbi.extensibility.visual.VisualUpdateOptions;
 import IVisual = powerbi.extensibility.visual.IVisual;