+++ templates/visuals/rvisual/src/visual.ts
@@ -25,7 +25,7 @@
 */
 
 "use strict";
-import powerbi from "powerbi-visuals-api";
+import powerbi from "@visualbi/powerbi-visuals-api";
 import VisualConstructorOptions = powerbi.extensibility.visual.VisualConstructorOptions;
 import VisualUpdateOptions = powerbi.extensibility.visual.VisualUpdateOptions;
 import IVisual = powerbi.extensibility.visual.IVisual;