+++ templates/visuals/table/src/visual.ts
@@ -31,7 +31,7 @@ import { select } from "d3-selection";
 import { transpose } from "d3-array";
 import { formatPrefix } from "d3-format";
 
-import powerbi from "powerbi-visuals-api";
+import powerbi from "@visualbi/powerbi-visuals-api";
 import VisualConstructorOptions = powerbi.extensibility.visual.VisualConstructorOptions;
 import VisualUpdateOptions = powerbi.extensibility.visual.VisualUpdateOptions;
 import IVisual = powerbi.extensibility.visual.IVisual;